Switch to using package traits for SubprocessFoundation
